+++ examples/compare_methods.py
@@ -5,7 +5,7 @@ import numpy as np
 import pybamm
 from scipy.fft import fft
 
-import pbeis
+import pybammeis
 
 # Set up
 model = pybamm.lithium_ion.SPM(options={"surface form": "differential"}, name="SPM")
@@ -58,7 +58,7 @@ methods = ["direct", "bicgstab", "prebicgstab"]
 impedances_freqs = []
 for method in methods:
     start_time = timer.time()
-    eis_sim = pbeis.EISSimulation(model, parameter_values=parameter_values)
+    eis_sim = pybammeis.EISSimulation(model, parameter_values=parameter_values)
     impedances_freq = eis_sim.solve(frequencies, method)
     end_time = timer.time()
     time_elapsed = end_time - start_time
@@ -67,9 +67,9 @@ for method in methods:
 
 # Compare
 _, ax = plt.subplots()
-ax = pbeis.nyquist_plot(impedances_time, ax=ax, label="Time", alpha=0.7)
+ax = pybammeis.nyquist_plot(impedances_time, ax=ax, label="Time", alpha=0.7)
 for i, method in enumerate(methods):
-    ax = pbeis.nyquist_plot(
+    ax = pybammeis.nyquist_plot(
         impedances_freqs[i], ax=ax, label=f"Frequency ({method})", alpha=0.7
     )
 ax.legend()
