+++ pybammeis/utils.py
